Remove lazy `Regexp` backend from extn
Remove the `Regexp::lazy` constructor and associated lazily-allocating
backend.

This change removes a 3p dep from `artichoke-backend` and reduces the
complexity of the `Regexp` implementation to prep for extraction to
`spinoso-regexp`.

`Regexp` is the oldest extn module and still uses the trait object based
multi-backend approach. All modern spinoso crates use source-level
polymorphism and expect artichoke-backend to select one concrete
implementation to embed into the VM. This means dyn-based indirection is
not a feasible approach.

Additionally, the lazy-allocating behavior relied on fallible
construction via a `once_cell` which tainted the implementation of
`Lazy` with lots of branching.

This removal will make extraction to a `spinoso-regexp` crate easier by
limiting the number of backends to implement (now only 3 instead of 6).

This is a moderate breaking change to the extn API, so bump
`artichoke-backend` to v0.6.0. The `Regexp::lazy` constructor is
replaced with a `TryFrom<Vec<u8>>` impl which constructs the `Source`
and `Config` arguments to `Regexp::new` in the same way `Lazy` did.