Add enums to use as results from `Kernel#load` and `Kernel#require`
In Ruby, `Kernel#load` returns `true` to indicate the file has been
successfully loaded and `Kernel#require` returns either `true` or
`false` depending on whether the file was previously present in the
interpreter's set of loaded features.

`LoadSources` in artichoke-core represented the return types in the `Ok`
variant of its load and require APIs as a boolean, but we can be more
expressive in Rust with enums.

These new enums give names to the outcomes and, for `load`, indicate
that only `true` is ever returned. These new enums have `into` methods
that convert into `bool` to return to the VM.