Remove more impls of `Borrow<[u8]>` in `spinoso-string`
Per the docs in `std` [0], for an impl of `Borrow<T>` to be correct:

> In particular `Eq`, `Ord` and `Hash` must be equivalent for borrowed
> and owned values: `x.borrow() == y.borrow()` should give the same
> result as `x == y`.

The encoded string subtypes are not guaranteed to uphold this invariant,
so remove the impls. This invariant _is_ meant to be upheld by
`EncodedString` and `String`; code has been moved and docs added to
reflect this intended contract.

[0] https://doc.rust-lang.org/std/borrow/trait.Borrow.html