Fix `extern "C" fn` signature for `mrb_str_aref`
This function takes three `mrb_value`s; it does not take an `mrb_int`.

Additionally, the only time this function is used in `vm.c`, it passes
an `UNDEF` `mrb_value` as the third argument. Properly detect this value
and transform it to `None` before delegating to the `trampoline::aref`
impl. mruby's implementation treats an `UNDEF` argument as if the first
argument is a `Range`.