Make `Value::inner` `pub(crate)` in artichoke-backend
I think this was fixed in the great de-rc/refcell event in #670 and was
never cleaned up in the issue tracker.

For crates external to `artichoke-backend` that define methods on the
mruby VM, like the `artichoke-backend` integration tests, `Value::into`
is available to get at the inner `sys::mrb_value` because `mrb_value`
impls `From<Value>`.

Fixes #234.
Fixes #251.