Remove obsolete `auto_import` scripts from artichoke-backend
The autoimport scripts were introduced in #117 as a mechanism for
fetching default gems from a host MRI Ruby installation for embedding in
Artichoke.

Since #579 and #629, this script is not run as part of `artichoke-backend`'s
`build.rs`. The sources it discovered have been checked in as vendored
sources in `extn/stdlib/vendor/gen`.

This script is legacy and unused, so delete it.