Refactor `Random` impl to rely on `spinoso-random`
The implementation of `Random` in `extn::core::random` differs from my
preferred way of implementing types in `extn`, which is like that of
`Time`: the Spinoso type is re-exported untouched and all argument
marshaling is done in the trampoline.

This commit reimplements `extn::core::random` so that all type
transforms, boxing, and unboxing happens in `random::trampoline`. This
allows `spinoso_random::Random` to be re-exported. The only remaining
code in `mod.rs` in `extn::core::random` is a bunch of trait
implementations to make that argument marshaling easier.

I also took the opportunity to factor out the seed truncation to a
single spot and make the logic a bit more clear with a transmute and
memcpy.