Revert "Address Security/CompoundHash lint violation in Range#hash impl"
This reverts commit 3186d1edb2dd17edd3a5ff395e838566f5f350b9.

Using an `Array` to collect the intermediate values to hash doesn't
actually work in the current implementation of `Array#hash`. The
implementation of `Array#hash` comes from `Kernel#hash` in mruby, which
defines the hash code in terms of object id.

In this case, we construct a new temporary object every time we compute
the hash for the `Range`, which means the hash code is not stable.