Add support for a `Range` argument to `String#byteslice`
Add support for `Range` arguments in `String#byteslice` of the form:

```ruby
s = "abc"
raise unless s.byteslice(1..2) == "bc"
```

As of this commit all `String#byteslice` specs pass except for two which
require the `begin` and `end` components of the `Range` to go through
the "implicitly convert to int" flow. Range parsing is still implemented
in mruby which does not do implicit conversions.

Getting the specs to pass required significant rework of the
`protect::is_range` function to distinguish between the "range out" case
and the "valid" case. For example:

```ruby
"".byteslice(-1..-1)
```

was previously squashing this range into the `None` case, which led to
`byteslice` blowing up further down the routine when attempting to fall
back to the integer argument path.

This change led to re-evaluating all other instances of range argument
handling and corrected the response when the "out of range" case is
encountered. The most significant change was in `Array#[]=` which raises
an exception in this case.

Fixes #1916.

Among the core specs:

Before:
Passed 3944, skipped 1063, not implemented 724, failed 4494 specs.

After:
Passed 4014, skipped 1065, not implemented 724, failed 4561 specs.