Add a NUL byte to the end of the allocation from `mrb_sym_str`
mruby assumes that symbols are stored in memory as a null terminated
`char*`s and creates "static" `RString`s from interned bytes that
point at this NUL terminated memory.

Sometimes mruby grabs the `RString` pointer directly from value
returned by this API call and assumes it is NUL terminated.

Add a 0 byte to the end of the `Vec` to force an allocation and then
set the length to the length of the byte content.

This was observed when printing out class names in ARGF specs.

`mrb_class_name` directly grabs the `RSTRING_PTR` from a String value:

https://github.com/artichoke/artichoke/blob/7bf3f033e020dd60f4b5fd5974dcb19af2617f6d/artichoke-backend/vendor/mruby/src/class.c#L2148

which searches the class path:

https://github.com/artichoke/artichoke/blob/7bf3f033e020dd60f4b5fd5974dcb19af2617f6d/artichoke-backend/vendor/mruby/src/class.c#L396

which is returns a class name from the symbol table:

https://github.com/artichoke/artichoke/blob/7bf3f033e020dd60f4b5fd5974dcb19af2617f6d/artichoke-backend/vendor/mruby/src/class.c#L2125