Implement String FFI constructors with `mrb_str_new_capa`
To maintain the invariant that all string byte buffers have a trailing
NUL byte, Artichoke has added logic to `mrb_str_new_capa` to always
allocate an extra byte and write a `0` to it.

To avoid repeating this error prone logic in multiple places, implement
`mrb_str_new` in terms of `mrb_str_new_capa` and a raw `memcpy`,
remembering to fixup the length in the `RString*`.

Implement `mrb_str_new_cstr` with `mrb_str_new` by finding the `ptr` and
`len` with Rust's `CStr`.