Delegate to `mrb_str_new` for `mrb_str_dup`
`mrb_str_dup` is one of the paths to return a `String` valuen from
`mrb_class_path`. This function is the same that was implicated in the
prior PR #1969, but the fix was incomplete and only targeted the
`mrb_sym_str` path.

https://github.com/artichoke/artichoke/blob/7bf3f033e020dd60f4b5fd5974dcb19af2617f6d/artichoke-backend/vendor/mruby/src/class.c#L2123-L2127

This branch has made improvements to `mrb_str_new_capa` and
`mrb_str_new` to ensure invariants around trailing NULs in byte buffer
allocations are maintained.

Rework the implementation of `mrb_str_dup` to delegate to `mrb_str_new`.