Implement custom wrappers around `Vec` in `spinoso-string`
Issues #1969 and #1975 fixed memory unsafety due to a failure to uphold
mruby's invariants that byte buffers stored in an `RString*` are always
NUL terminated.

PRs #1969 and #1975 made tactical fixes at critical "FFI" portions of
the `MRB_API` surface that are likely to be used directly by the mruby
VM, but these tactical fixes are incomplete.

This commit adds two custom `Vec` types to `spinoso-string`. These
crate-internal `Vec`s will be used by all encoded string types and can
be specified at compile time with a Cargo feature.

One variant added in this commit is a transparent wrapper around `Vec`;
it does not change the semantics of any APIs or influence the way the
inner `alloc::vec::Vec` allocates.

The various encoded string variants in `spinoso-string` are updated to
use this new struct as their internal byte buffer.

This commit also adds a second variant that on construction and mutation
ensures that the buffer is NUL terminated in the spare capacity of the
`Vec` by using the `Vec::spare_capacity_mut` and `Vec::reserve_exact` APIs.
The NUL terminator will be present in the uninitialized memory that is part
of the `Vec`'s allocation and follows the `len` of the `Vec`.

Several `const` APIs have been made non-`const` to allow for future
variants to allocate (for example, in `String::new`). The crate version
has been bumped to 0.19.0.