Upgrade `rustyline` to 10.0.0
`rustyline` 10.0.0 includes a breaking change where `Editor::new` now
returns result. The REPL propagates this error as an unhandled readline
error, which will exit the REPL with an error message.

This is a handcrafted version of @dependabot's PR in #1999 to address
this breaking change.

`rustyline` 10.0.0 includes several commits (a few of which I authored)
that reduce the number of deps pulled in a bunch:

- https://github.com/kkawakam/rustyline/pull/613
- https://github.com/kkawakam/rustyline/pull/614
- https://github.com/kkawakam/rustyline/pull/615
- https://github.com/kkawakam/rustyline/pull/623