Enable dhat heap profiling in spec-runner with a Cargo feature
This commit contains the harness used to identify unnecessary
allocations and improve performance in #2041.

The `dhat` crate is added to the spec-runner gated by a cargo feature as
shown in the `dhat` docs: https://docs.rs/dhat/0.3.0/dhat/index.html.

The spec-runner can be invoked with these flags to produce a dhat heap
profile:

```console
$ cargo run --release --bin spec-runner -q --features dhat-heap -- enforced-specs.toml
```

This will run the spec harness with the given config and produce a
`dhat-heap.json` file. That file can be loaded into a web viewer [0] to
browse heavy allocation sites by the lowest frames on the stack.

[0] https://nnethercote.github.io/dh_view/dh_view.html

This Twitter thread is an introduction to dhat-rs:

https://twitter.com/nnethercote/status/1336132646020104193