Implement `Array#<<` in Rust
After changing the visibility of the extn trampolines,
`extn::core::array::trampoline::push` emitted a warning for being
unused. Upon inspection, `Array#<<` is not implemented in Rust. This
function was instead implemented in pure Ruby, which delegates to
`Array#push`. `Array#push` delegates to `Array#concat`.

Using `Array#concat` to implement `Array#<<` is incorrect since the
given element is meant to be pushed directly instead of flattened if it
is an `Array` like `Array#concat` does.

Fixes #2114.