Parallelize `cc` and `bindgen` steps in backend build script
As of trunk c5082eadc0e4052b5655b33b89efb7f8c1b6d0c2, a run of
`cargo +nightly build --timings` gives a 26.8s build on my machine. The
long pole for this build is the `artichoke-backend` build script, which
takes over 14 seconds to run.

`artichoke-backend` takes an additional 6 seconds to compile on top of
the build script, so improvements we can make to speed up the build
script have linear improvements on build time.

Rust 1.63.0, which Artichoke uses as its Rust toolchain as of #2076,
introduces scoped threads to `std`. Scoped threads allow for parallelism
which is scoped to a stack frame. This allows us to treat parallel
execution similarly to an imperative style.

The `artichoke-backend` build script has two expensive halves:

- Compile mruby and mrbsys C sources into a static library using `cc`.
- Generate Rust code from mruby C headers using `bindgen`.

This patch spawns two threads using `std::thread::scope` to run these
tasks in parallel. This single change makes the entire build of
Artichoke from a clean state 35% faster. On my machine, build times
dropped from 26.8s to 17.2s. The long pole of the build script dropped
from 14.36s to 6.26s.

Timing reports from the compiler are attached to the PR associated with
this patch.