Enable `cc` crate to build C sources in parallel
Enable the `parallel` feature on `cc` in artichoke-backend's build
dependencies. This feature pulls in the `jobserver` crate and allows
`cc` to drive parallelism by integrating with `cargo`'s jobserver.

This change reduces two long poles of the build:

- `onig_sys` build script: 5.8s -> 1.9s
- `artichoke-backend` build script: 6.9s -> 2.1s

After this change, neither of these crates are the long pole. After this
change the next two longest poles are:

- Deep dependency tree from `onig_sys` build script -> `onig_sys` ->
  `onig` -> `spinoso-regexp`.
- `tzdb` (5.4s) -> `spinoso-time`.

`artichoke-backend` itself also is a chunky crate, taking 6.5s to
compile.

Followup to #2117 with some more goodness from `cargo build --timings`.