Remove chrono backend from spinoso-time
All of the recent investment in `spinoso-time` has been poured into the
`tz-rs` backend and we've got a high-quality time implementation with
the `tzrs` feature.

The `chrono` backend has atrophied since it was introduced. The API has
diverged from the standard set by the `tz-rs` backend (many APIs which
are fallible in the `tz-rs` backend are infallible in the `chrono`
backend). Full support for the offset API is not something that is
easily achievable with `chrono`.

`chrono` itself is difficult to work with. For example, many APIs panic
on out of range inputs. The API is overly generic which forces
`spinoso-time` to deconstruct a `DateTime` into its raw parts and use
`chrono` types as "views" into the calendar in each accessor method.

`chrono` with time zone support is also a heavy dependency. Removing
this backend results in 13 crates being dropped from `Cargo.lock`.

Especially given the good collaboration we've had with @x-hgg-x (`tz-rs`)
and @Kijewski (`tzdb`, `iana-time-zone`), I don't see a compelling reason
to continue maintaining the `chrono` backend in `spinoso-time`.