Replace `bstr/std` with `scolapasta-path` in `spinoso-env`
The "system env" backend requires converting between bytes and platform
strings. Previously this was done with routines which require the
`bstr/std` feature. This feature is a bit heavy though ans the
`OsString` specific bits are available in `scolapasta-path`.

This commit also includes changes to get `spinoso-env` to build and test
with no default features.

Followup to https://github.com/artichoke/artichoke/pull/2173.