Remove an allocation when computing a fixed time zone offset
All valid timezone fixed time zone offsets result in an offset string of
the format +HHMM or -HHMM. Both forms are exactly 5 bytes. We can use a
stack-allocated buffer to write the bytes into and avoid allocating a
`String`.

Exhaustive tests are added to assert the invariant that the buffer is
exactly the right size.