Limit downcasing work for `String` Time month args
All valid string arguments to `Time.new` for the month are 3 ASCII
characters. In the event that a very long string is passed, the existing
implementation will do a lot of work downcasing the string that is
ultimately discarded.

Short circuit this by extracting the first 3 bytes from the captured
`Vec` and only downcasing those.