Close interpreter if there are errors during init
This updates the REPL and interpreter boot to prevent memory leaks if
there are errors or exceptions during interpreer boot and initialization.

This commit follows the same approach in:

- https://github.com/artichoke/artichoke/pull/1032

The REPL regressed in:

- https://github.com/artichoke/artichoke/pull/2423

As far as I can tell, interpreter boot in `artichoke-backend` has always
had these leak problems on error.