Remove dependency on `directories` and `dirs-sys`
`directories` and `dirs-sys` are no longer acceptable dependencies due
to licensing issues. See:

- https://github.com/artichoke/artichoke/pull/2543

The recent substitution of `directories::ProjectDirs` for
`artichoke-repl-history` removed the last use of the `directories`
crate. This commit removes it as an explicit dependency.

This change drops 15 deps from `Cargo.lock` including the `thiserror`
proc macro, Redox support crates which are not necessary since Redox
became a unix target in Rustc, and duplicate versions of `windows-sys`.