Significantly simplify and improve performance of Utf8String::index
Upon review of the work @choznerol and @lopopolo did, I noticed that the
structure of `Utf8String::index` could mirror that of the ASCII and
binary implementations.

The ASCII and binary implementations did not look applicable at first
pass because they did not have the initial step to compute the offset
char length (they skip this because all characters are 1 byte).

Using `Utf8String::get_char_slice` though, we can convert the offset
char count into a byte position, use `BStr::find` as we previously were,
and then use the newly extracted `char_len` function on the matched
portion of the string to find the char equivalent of that index.
Converting the provided offset and found index to "char counts" allows
the same basic structure of the routine to remain the same.

This is significantly cheaper because it allows us to delegate most of
the char counting work to `bytecount`, which is SIMD optimized and
should be much faster than doing a partial decode one character at a
time with `bstr`.