Use `std::sync::OnceLock` in spinoso-time
Drop the dependency on `once_cell` for the newly stabilized in Rust
1.70.0 `OnceLock`. `OnceLock` is the `std` version of `once_cell`'s
`once_cell::sync::OnceCell` type.

This hurts readability a bit because the definition of the static is
separate from the construction of the lazy value, but `std`'s equivalent
of the `Lazy` type is still unstable.

Prefer the readability hit to drop a dependency.