Split UTF-8 string variant into owned and borrowed types
`spinoso-string` implements the various encodings it supports with one
type per encoding. For the UTF-8 encoding, the type is `Utf8String`.

`Utf8String` is conceptually similar to `Vec<u8>` and `String` in `std`.
Both of these types deref to a slice reference counterpart (`&[u8]` and
`&str`). All methods which do not require growing the backing vector (or
dynamic allocation more generally) are implemented on the slice ref part
of the type pair.

All of the encoding variants in `spinoso-string` do not have this
flexible API. This presents challenges when wanting to perform
encoding-oriented operations on a borrowed byte slice. A recent example
of this was the refactoring required to get UTF-8 char len calculation
shared among various parts of the code:

- https://github.com/artichoke/artichoke/commit/67726d1a2f52f83830e7b2374e07c2b4337440ad
- https://github.com/artichoke/artichoke/pull/2554

This commit experiments with breaking apart `Utf8String` into an owned
and borrowed pair of types: `Utf8String` and `&Utf8Str`/`&mut Utf8Str`.
The borrowed types are heavily inspired by `bstr::BStr`.

The benefits of the reference types that associate a byte slice to an
encoding mostly arise from avoiding the need to allocate an owned
`Utf8String`. I've long viewed a `String`'s encoding as a "view" or
"cursor" over the underlying byte content. Being able to interrogate an
arbitrary byte slice via a specific encoding makes the API much more
powerful.

Eventually I would like to extract APIs like `Utf8String`/`Utf8Str` to
their own crates in the workspace. The code is sufficiently complicated
and there is _a lot_ of it. Doing so would allow exposing the various
encodings in a `no-std::no-alloc` configuration.

This commit contains one breaking change which led to the version bump
in `spinoso-string` (removed `impl Extend<&'a mut u8>`), but
surprisingly little of the crate had to change outside of the UTF-8
internals.

I still consider this approach experimental and am looking for feedback
on the design.

One WIP in the design: ASCII and Binary encoded strings will be able to
implement `reverse()` on their reference type. Due to the complexity in
a character-wise reversal of the conventionally UTF-8 byte contents in
`Utf8String`, allocation is required so this API still lives on
`Utf8String`.

A benefit of the borrowed type design is (at least for me) it makes it
more obvious how a crate like `encoding_rs` could be used to support
many more encodings than Artichoke does today.

As an aside, sorry for the massive commit. I ended up working on #2589
_after_ I made these changes and I couldn't resolve the merge conflicts.