Improve the readability of dependency manifests in `Cargo.toml`
In the same vein of quality of life improvements from #2593, this commit
fixes something that has been bothering me for a while.

Dependency declarations in the `Cargo.toml` manifests in this repository
have been growing unwieldy. Especially for workspace crates which are
depended on by `path` attribute, the inline table can extend far to the
right of my terminal width. Some crates, like `scolapasta-string-escape`
are impacted badly by this because the crate name is long and it is
often used without default features.

With this change, _all_ dependencies which include an inline table (i.e.
anything other than a bare version specifier string) are expanded into a
dedicated table.

This turns something like this:

```toml
scolapasta-string-escape = { version = "0.3.0", path = "../scolapasta-string-escape", optional = true, default-features = false }
```

into this:

```toml
[dependencies.scolapasta-string-escape]
version = "0.3.0"
path = "../scolapasta-string-escape"
optional = true
default-features = false
```

This also gave the opportunity to clean up all of the dependency
declarations for consistency. In the table, keys should appear in the
following order:

- version
- path
- optional
- default-features
- features

Top-level tables such as `[dependencies]` should appear in `Cargo.toml`
even if they are empty.

This change is made in all workspaces.