Omit return value inspection in `airb`
In `irb` v1.10.0 (included in Ruby v3.3.0), users can omit return value
inspection (printing out the contents of `#inspect` for an expression
after evaluating it) by ending the input with the `;` character.

This removes a workaround I personally used of suffixing lines with
`; nil`.

See: https://railsatscale.com/2023-12-19-irb-for-ruby-3-3/#omitting-return-value-inspection-with-