Only drop an `Array` during free if the `Array` is initialized
In Ruby, objects are normally created by calling `<Class>.new`, but
under the covers this does a Smalltalk-style sequence of calling:

```
instance = <Class>.allocate
instance.initialize
```

`Object::allocate` is directly callable by application code, and there
are Ruby Specs which exercise this behavior. This permits a state where
an `RArray*` object is created and live on the GC heap but does not
contain an array buffer.

When upgrading to Rust 1.82.0, new pointer operation well-formedness
assertions from `core` in debug builds were firing in `mrb_ary_artichoke_free`
when attempting to construct a `Vec` with a null pointer.

To address this source of UB, first check whether the pointer in the
`RArray*` is null by parsing it with `std::ptr::NonNull`. If the `RArray*`
contains a null pointer, that means we have an uninitialized `Array` and
can safely skip dropping a backing buffer.