Add failing test for safe garbage collection of uninitialized `Array`
An `Array` which has been allocated via `Array.allocate`, but not
initialized via `Array#initialize`, will not have a backing buffer and
instead will store a null pointer in the `RArray*` representation.

This will result in a debug assert in new Rust (and is currently UB)
when constructing the `spinso-array`'s underlying `Vec` on drop during a
garbage collection cycle.

On Rust 1.82.0, this test fails with a debug assertion in alloc's
`RawVec` internals when consturcting a `NonNull` from the pointer given
to `Vec::from_raw_parts`.