Add failing tests for safe read and write operations on uninitialized `Array`
An `Array` which has been allocated via `Array.allocate`, but not
initialized via `Array#initialize`, will not have a backing buffer and
instead will store a null pointer in the `RArray*` representation.

This will result in a debug assert in new Rust (and is currently UB)
when constructing the `spinso-array`'s underlying `Vec` when unpacking
the `RArray*` from the `mrb_value` in the array trampoline module.

On Rust 1.82.0, this test fails with a debug assertion in alloc's
`RawVec` internals when consturcting a `NonNull` from the pointer given
to `Vec::from_raw_parts`.