Only drop a `String` during free if the `String` is initialized
In Ruby, objects are normally created by calling `<Class>.new`, but
under the covers this does a Smalltalk-style sequence of calling:

```
instance = <Class>.allocate
instance.initialize
```

`Object::allocate` is directly callable by application code, and there
are Ruby Specs which exercise this behavior. This permits a state where
an `RString*` object is created and live on the GC heap but does not
contain a byte buffer.

When upgrading to Rust 1.82.0, new pointer operation well-formedness
assertions from `core` in debug builds were firing in `mrb_gc_free_str`
when attempting to construct a `Vec` with a null pointer.

To address this source of UB, first check whether the pointer in the
`RString*` is null by parsing it with `std::ptr::NonNull`. If the `RString*`
contains a null pointer, that means we have an uninitialized `String` and
can safely skip dropping a backing buffer.