Add failing tests for safe read and write operations on uninitialized `String`
A `String` which has been allocated via `String.allocate`, but not
initialized via `String#initialize`, will not have a backing buffer and
instead will store a null pointer in the `RString*` representation.

This will result in a debug assert in new Rust (and is currently UB)
when constructing the `spinso-string`'s underlying `Vec` when unpacking
the `RString*` from the `mrb_value` in the string trampoline module.

On Rust 1.82.0, this test fails with a debug assertion in alloc's
`RawVec` internals when consturcting a `NonNull` from the pointer given
to `Vec::from_raw_parts`.