Ensure allocated but uninitialized `String` have a backing byte buffer
In Ruby, objects are normally created by calling `<Class>.new`, but
under the covers this does a Smalltalk-style sequence of calling:

```
instance = <Class>.allocate
instance.initialize
```

`Object::allocate` is directly callable by application code, and there
are Ruby Specs which exercise this behavior. This permits a state where
an `RString*` object is created and live on the GC heap but does not
contain a string buffer.

When upgrading to Rust 1.82.0, new pointer operation well-formedness
assertions from `core` in debug builds were firing in
`<String as BoxUnboxVmValue>::unbox_from_value` when attempting to
construct a `Vec` with a null pointer.

On reads, CRuby will treat an allocated but uninitialized `String` as
empty. On writes, CRuby will prepare the string to be modified, which
among other things ensures there is a backing buffer in the string sized
with enough capacity to accept the writes.

To address this source of UB, first check whether the pointer in the
`RString*` is null by parsing it with `std::ptr::NonNull`. If the `RString*`
contains a null pointer, that means we have an uninitialized `String` and
can swap in an empty buffer with `String::new`. This will not allocate,
but does prepare the unboxed `String` for reading and writing by swapping
in a valid Rust buffer.

Additionally, this will leave the `String` uninitialized on reads because
the `String` will not be reboxed back into the source `RString*`.