Fix memory leak in `Array#initialize`
Ruby allows calling `#initialize` on any object, including ones that are
already initialized. Previously, the trampoline code for `Array` did not
take this into account and did not try to drop an existing buffer if
present. This caused a memory leak when the `Array` internals were then
overwritten with an empty buffer.

This logic mirrors what already exists for `String`.