Move conversion method initializers into interpreter state
Previously, the conversion method table used a static initializer to
also intern symbols into an Artichoke interpreter, which means using
multiple interpreters in a single process was likely subtly incorrect.

Extract out the conversion method table, initialization, and searching
routines to dedicated types in a new `mezzaluna-conversion-methods`
crate with docs and tests.

Additionally, add a new `ConvMethods` container to the Artichoke
`State` and new top-level methods on the interpreter to permit sharing
the intaglio interner.

Fixes #2783.